Moved controller function for grant value out to a new controller
