+++ src/main/resources/routes
@@ -55,10 +55,10 @@ GET         /manage/opportunity/:id/title/edit                        controller
 GET         /manage/opportunity/:id/section/:num/questions            controllers.manage.OpportunityController.viewQuestions(id: OpportunityId, num:Int)
 
 
-GET         /manage/opportunity/:id/grantvalue                        controllers.manage.OpportunityController.viewGrantValue(id:OpportunityId)
-POST        /manage/opportunity/:id/grantvalue                        controllers.manage.OpportunityController.saveGrantValue(id:OpportunityId)
-GET         /manage/opportunity/:id/grantvalue/edit                   controllers.manage.OpportunityController.editGrantValue(id:OpportunityId)
-GET         /manage/opportunity/:id/grantvalue/preview                controllers.manage.OpportunityController.previewGrantValue(id:OpportunityId)
+GET         /manage/opportunity/:id/grantvalue                        controllers.manage.GrantValueController.viewGrantValue(id:OpportunityId)
+POST        /manage/opportunity/:id/grantvalue                        controllers.manage.GrantValueController.saveGrantValue(id:OpportunityId)
+GET         /manage/opportunity/:id/grantvalue/edit                   controllers.manage.GrantValueController.editGrantValue(id:OpportunityId)
+GET         /manage/opportunity/:id/grantvalue/preview                controllers.manage.GrantValueController.previewGrantValue(id:OpportunityId)
 
 GET         /manage/opportunity/:id/section/:num                      controllers.manage.OpportunityController.viewSection(id:OpportunityId, num:Int)
 POST        /manage/opportunity/:id/section/:num                      controllers.manage.OpportunityController.saveSection(id:OpportunityId, num:Int)
