+++ src/main/twirl/views/manage/opportunityOverview.scala.html
@@ -77,7 +77,7 @@ backLink=Some(BackLink("Opportunity library", controllers.manage.routes.Opportun
                                 <tr>
                                     <td>
                                         3. <a id="op-setup-link3"
-                                              href="@controllers.manage.routes.OpportunityController.viewGrantValue(opportunity.id)"
+                                              href="@controllers.manage.routes.GrantValueController.viewGrantValue(opportunity.id)"
                                             > Maximum grant value</a>
                                     </td>
                                 </tr>
